unify version parsing in Version.py
