+++ setup.py
@@ -5,6 +5,7 @@ from setuptools import setup, Extension
 from python.lib.moresetuptools import *
 # function to generate gen/*.{c,h}
 from python.lib.gen_external import generate_external, header, output_path
+from Version import get_aubio_version, get_aubio_pyversion
 
 __version__ = get_aubio_version()
 __pip_version__ = get_aubio_pyversion()
